+++ packages/bunshine/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v3.2.3 - Jan 13, 2024
+
+- `c.sse()` to support multi-line messages
+- Accommodate Bun 1.1.43 bugfix to HEAD content-length header
+- Avoid etags and compression when body is a stream
+- Add `maxSize` option to compression middleware (default 2GB)
+- Add LRU cache for file-based mime detection
+
 ## v3.2.2 - Dec 7, 2024
 
 - Fix mime-type on css